Clock Setter
This Addon is really for a personal project but I decided to post it publicly as it might have useful patterns when learning.

Simple input plugin that asks for "." delimited with the syntax of `[HH.][MM.]SS`; and converts that into int8 attribute data for my Art-Net Clock

Not really helpful if you don't have the Clock but shows how to receive input and set attributes to a specific type value.

Nothing is finalized other then testing that it with the correct input causes the correct output.